toyRent/mypro/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.2
idna==3.7
mysqlclient==2.2.4
razorpay==1.4.2
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
